Fix non jit
